token realizado

+++ src/controllers/usuarios.controller.js
@@ -1,39 +1,67 @@
 import { Auxiliar } from '../models/Auxiliar.js';
 import { Usuario } from '../models/Usuario.js';
-import {v4 as uuidv4 } from 'uuid'
 
-export const getUsuarios = async (req, res) => {
-    try {
-        const usuarios = await Usuario.findAll();
-        res.json(usuarios);
-    } catch (error) {
-        return res.status(500).json({ message: error.message });
-    }
-}
+import bcrypt from 'bcrypt';
+import dotenv from 'dotenv';
+import jwt from 'jsonwebtoken';
+
+dotenv.config(); 
+
+// export const getUsuarios = async (req, res) => {
+//     try {
+//         const usuarios = await Usuario.findAll();
+//         res.json(usuarios);
+//     } catch (error) {
+//         return res.status(500).json({ message: error.message });
+//     }
+// };
 
 export const getUsuario = async (req, res) => {
     try {
-        const { id_usuario } = req.params;
-        const usuario = await Usuario.findOne({ where: { id_usuario } });
-        if (!usuario) return res.status(404).json({ message: "El usuario no existe" });
-        res.json(usuario);
-    } catch (error) {
-        return res.status(500).json({ message: error.message });
-    }
-}
+        const token = req.headers['x-access-token']
 
-export const createUsuario = async (req, res) => {
-    const { nombre_usuario, contrasenia_usuario, email_usuario } = req.body;
-    try {
-        const newUsuario = await Usuario.create({ nombre_usuario, contrasenia_usuario, email_usuario });
-        res.json(newUsuario);
+        if(!token){
+            return res.status(401).json({
+                auth: false,
+                message: 'No token provided'
+            })
+        }
+
+        const decoded = jwt.verify(token, process.env.JWT_SECRET)
+
+        const user = await Usuario.findByPk(
+            decoded.id, 
+            {attributes: ['id_usuario', 'nombre_usuario', 'email_usuario', 'tipo_usuario']} 
+        );
+
+        res.json(user);
     } catch (error) {
         return res.status(500).json({ message: error.message });
     }
-}
+};
+
+// export const createUsuario = async (req, res) => {
+//     try {
+//         const { nombre_usuario, contrasenia_usuario, email_usuario, tipo_usuario } = req.body;
+//         const newUsuario = await Usuario.create({ nombre_usuario, contrasenia_usuario, email_usuario, tipo_usuario });
+//         res.json(newUsuario);
+//     } catch (error) {
+//         return res.status(500).json({ message: error.message });
+//     }
+// };
 
 export const updateUsuario = async (req, res) => {
     try {
+
+        const token = req.headers['x-access-token']
+
+        if(!token){
+            return res.status(401).json({
+                auth: false,
+                message: 'No token provided'
+            })
+        }
+
         const { id_usuario } = req.params;
         const { nombre_usuario, contrasenia_usuario, email_usuario } = req.body;
 
@@ -50,56 +78,53 @@ export const updateUsuario = async (req, res) => {
     } catch (error) {
         return res.status(500).json({ message: error.message });
     }
-}
+};
 
 export const deleteUsuario = async (req, res) => {
     try {
+
+        const token = req.headers['x-access-token']
+
+        if(!token){
+            return res.status(401).json({
+                auth: false,
+                message: 'No token provided'
+            })
+        }
+
         const { id_usuario } = req.params;
         await Usuario.destroy({ where: { id_usuario } });
         res.sendStatus(204);
     } catch (error) {
         return res.status(500).json({ message: error.message });
     }
-}
+};
 
+export const loginUsuario = async (req, res) => {
+    const { nombre_usuario, contrasenia_usuario } = req.body;
 
-// export const validarUsuario = async (req, res) => {
-//     try {
-//         const { codsiss, contrasenia_usuario } = req.body;
+    try {
+        const usuario = await Usuario.findOne({ where: { nombre_usuario } });
 
-//         const usuario = await Usuario.findOne({
-//             where: {
-//                 codsiss,
-//                 contrasenia_usuario,
-//             },
-//         });
+        if (!usuario) {
+            return res.status(404).json({ message: "El usuario no existe" });
+        }
 
-//         if (!usuario) {
-//             return res.json({ estado: 'failed' });
-//         }
-//         const token = uuidv4();
+        const isMatch = await bcrypt.compare(contrasenia_usuario, usuario.contrasenia_usuario);
 
-//         const usuarioFormateado = {
-//             estado: 'successful',
-//             usuario: {
-//                 id_usuario: usuario.id_usuario,
-//                 token,
-//                 nombre_usuario: usuario.nombre_usuario,
-//                 email_usuario: usuario.email_usuario,
-//                 tipo_usuario: usuario.tipo_usuario,
-//                 codsiss: usuario.codsiss,
-//                 disponible: usuario.disponible,
-//                 ci_usuario: usuario.ci_usuario,
-//                 foto_usuario: usuario.foto_usuario,
-//             },
-//         };
+        if (!isMatch) {
+            return res.status(404).json({ message: "Password incorrecto" });
+        }
 
-//         res.json(usuarioFormateado);
-//     } catch (error) {
-//         return res.status(500).json({ message: error.message });
-//     }
-// };
+        const token = jwt.sign(
+            { id: usuario.id_usuario },
+            process.env.JWT_SECRET,
+            { expiresIn: '1h' }
+        );
 
-
-
-
+        res.json({ auth: true, token });
+    } catch (error) {
+        console.error('Error logging in user:', error);
+        res.status(500).json({ message: 'An error occurred during login' });
+    }
+};
